+++ packages/react-native-editor/CHANGELOG.md
@@ -10,6 +10,7 @@ For each user feature we should also add a importance categorization label  to i
 -->
 
 ## Unreleased
+-   [internal] Fix Inserter items list filtering [#62334]>>>>>>> trunk
 -   [*] Prevent hiding the keyboard when creating new list items [#62446]
 
 ## 1.120.0